Fix grammar issues and standardize LICENSE heading (#802)
This pull request focuses on improving the documentation by fixing
typos, grammatical errors, and standardizing terminology in markdown
files. The changes enhance clarity and consistency across the
documentation.

### Documentation Improvements:

* Corrected typographical errors in the `README.md`, such as "cells's"
to "cells'" and "we're looking forward to hear" to "we're looking
forward to hearing" for better grammar and readability.
* Standardized terminology, including changing "LICENCE" to "LICENSE"
for consistency with common usage. [
* Improved phrasing in `README.md` and
`internal/renderer/testdata/simple.md`, such as "help and information to
parameters" to "help and information on parameters" and "best effort
approach to extracts" to "best-effort approach to extract" for
grammatical correctness.

Signed-off-by: Arya Pathak <arya20j4@gmail.com>